requirements.txt
Flask
google-genai
python-dotenv
psycopg2-binary
markupsafe
requests
gunicorn
python-json-logger
Werkzeug
markdown2
markupsafe